new_requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
beartype==0.14.1
bitrix24-rest==1.1.1
Brotli==1.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cryptography==40.0.2
cssselect2==0.7.0
Django==4.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
docx2pdf==0.1.8
et-xmlfile==1.1.0
fast-bitrix24==1.5.16
fonttools==4.47.0
forex-python==1.8
frozenlist==1.4.0
future==0.18.3
html5lib==1.1
icontract==2.6.2
idna==3.4
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
more-itertools==9.1.0
multidict==6.0.4
numpy==1.25.2
numpy-financial==1.0.0
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.1.0
pdf417gen==0.7.1
pdfkit==1.0.0
phonenumbers==8.13.27
Pillow==10.0.0
psycopg2-binary==2.9.6
pycparser==2.21
pydyf==0.8.0
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==2.6.0
pypdf==3.15.5
PyPDF2==3.0.1
pyphen==0.14.0
pypng==0.20220715.0
python-barcode==0.15.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tabulate==0.9.0
tinycss2==1.2.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==1.26.15
weasyprint==60.2
webencodings==0.5.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.11
yarl==1.9.2
zopfli==0.2.3
